+++ TicketManagement/src/main/java/com/railway/TicketManagement/entities/Ticket.java
@@ -15,7 +15,7 @@ public class Ticket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ticket_id")
-    private Long ticketId;
+    private Integer ticketId;
 
     @ManyToOne
     @JoinColumn(name = "booking_id")
